requirements.txt
cryptography>=41
psptool>=2.6
asn1tools>=0.166
pycryptodome>=3.18
